김태훈 : c1280 IdleType 추가 및 BaseActor Reset 가상함수 추가

+++ DirectX_3DTeamPortfolio/GameEngineContents/Player.cpp
@@ -1333,7 +1333,7 @@ bool Player::GetHit(const HitParameter& _Para /*= HitParameter()*/)
 	Stat.AddPoise(-Stiffness);
 	Stat.AddHp(-AttackerAtt);
 
-	Hp = Stat.GetHp();
+	Hp = (float)Stat.GetHp();
 
 
 	return true;
